test: reduce code duplication after splitting test files in aws-lambda/test/integration_test
Keeping the *_test.js files separate per Lambda API flavor, but
factoring out the actual test code in one single test definition file
should give us the desired improvements in test distribution on CI while
avoiding excessive test code duplication.
